requirements.txt
streamlit 
PyMuPDF 
langchain 
sentence-transformers
faiss-cpu
python-dotenv 
langchain-community  
pypdf 
langchain-cohere